+++ src/main/java/org/influxdb/InfluxDB.java
@@ -297,6 +297,25 @@ public interface InfluxDB {
                     final ConsistencyLevel consistency, final String records);
 
   /**
+   * Write a set of Points to the influxdb database with the string records.
+   *
+   * @see <a href="https://github.com/influxdb/influxdb/pull/2696">2696</a>
+   *
+   * @param database
+   *          the name of the database to write
+   * @param retentionPolicy
+   *          the retentionPolicy to use
+   * @param consistency
+   *          the ConsistencyLevel to use
+   * @param precision
+   *          the time precision to use
+   * @param records
+   *            the points in the correct lineprotocol.
+   */
+  public void write(final String database, final String retentionPolicy,
+          final ConsistencyLevel consistency, final TimeUnit precision, final String records);
+
+  /**
    * Write a set of Points to the influxdb database with the list of string records.
    *
    * @see <a href="https://github.com/influxdb/influxdb/pull/2696">2696</a>
@@ -312,6 +331,25 @@ public interface InfluxDB {
    */
   public void write(final String database, final String retentionPolicy,
                     final ConsistencyLevel consistency, final List<String> records);
+
+  /**
+   * Write a set of Points to the influxdb database with the list of string records.
+   *
+   * @see <a href="https://github.com/influxdb/influxdb/pull/2696">2696</a>
+   *
+   * @param database
+   *          the name of the database to write
+   * @param retentionPolicy
+   *          the retentionPolicy to use
+   * @param consistency
+   *          the ConsistencyLevel to use
+   * @param precision
+   *          the time precision to use
+   * @param records
+   *          the List of points in the correct lineprotocol.
+   */
+  public void write(final String database, final String retentionPolicy,
+          final ConsistencyLevel consistency, final TimeUnit precision, final List<String> records);
 
   /**
    * Write a set of Points to the influxdb database with the string records through UDP.