+++ src/main/java/org/influxdb/dto/BatchPoints.java
@@ -5,6 +5,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.concurrent.TimeUnit;
 import java.util.TreeMap;
 
 import org.influxdb.InfluxDB.ConsistencyLevel;
@@ -25,6 +26,7 @@ public class BatchPoints {
   private Map<String, String> tags;
   private List<Point> points;
   private ConsistencyLevel consistency;
+  private TimeUnit precision;
 
   BatchPoints() {
     // Only visible in the Builder
@@ -50,6 +52,7 @@ public class BatchPoints {
     private final Map<String, String> tags = new TreeMap<>();
     private final List<Point> points = new ArrayList<>();
     private ConsistencyLevel consistency;
+    private TimeUnit precision;
 
     /**
      * @param database
@@ -117,6 +120,16 @@ public class BatchPoints {
     }
 
     /**
+     * Set the time precision to use for the whole batch. If unspecified, will default to {@link TimeUnit#NANOSECONDS}
+     * @param precision
+     * @return the Builder instance
+     */
+    public Builder precision(final TimeUnit precision) {
+      this.precision = precision;
+      return this;
+    }
+
+    /**
      * Create a new BatchPoints instance.
      *
      * @return the created BatchPoints.
@@ -135,6 +148,10 @@ public class BatchPoints {
         this.consistency = ConsistencyLevel.ONE;
       }
       batchPoints.setConsistency(this.consistency);
+      if (null == this.precision) {
+        this.precision = TimeUnit.NANOSECONDS;
+      }
+      batchPoints.setPrecision(this.precision);
       return batchPoints;
     }
   }
@@ -182,6 +199,20 @@ public class BatchPoints {
    */
   void setPoints(final List<Point> points) {
     this.points = points;
+  }
+
+  /**
+   * @return the time precision unit
+   */
+  public TimeUnit getPrecision() {
+    return precision;
+  }
+
+  /**
+   * @param precision the time precision to set for the batch points
+   */
+  void setPrecision(final TimeUnit precision) {
+    this.precision = precision;
   }
 
   /**
@@ -239,12 +270,13 @@ public class BatchPoints {
             && Objects.equals(retentionPolicy, that.retentionPolicy)
             && Objects.equals(tags, that.tags)
             && Objects.equals(points, that.points)
-            && consistency == that.consistency;
+            && consistency == that.consistency
+            && precision == that.precision;
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(database, retentionPolicy, tags, points, consistency);
+    return Objects.hash(database, retentionPolicy, tags, points, consistency, precision);
   }
 
   /**
@@ -261,6 +293,8 @@ public class BatchPoints {
            .append(this.consistency)
            .append(", tags=")
            .append(this.tags)
+           .append(", precision=")
+           .append(this.precision)
            .append(", points=")
            .append(this.points)
            .append("]");
@@ -275,8 +309,9 @@ public class BatchPoints {
    */
   public String lineProtocol() {
     StringBuilder sb = new StringBuilder();
+
     for (Point point : this.points) {
-      sb.append(point.lineProtocol()).append("\n");
+      sb.append(point.lineProtocol(this.precision)).append("\n");
     }
     return sb.toString();
   }