+++ src/test/java/org/influxdb/InfluxDBTest.java
@@ -17,6 +17,9 @@ import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
 
 import java.io.IOException;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -53,7 +56,7 @@ public class InfluxDBTest {
 		this.influxDB = TestUtils.connectToInfluxDB();
 		this.influxDB.createDatabase(UDP_DATABASE);
 	}
-	
+
 	/**
 	 * delete UDP database after all tests end.
 	 */
@@ -450,6 +453,158 @@ public class InfluxDBTest {
         Assertions.assertEquals("test3", result.getResults().get(0).getSeries().get(2).getTags().get("atag"));
         this.influxDB.deleteDatabase(dbName);
     }
+
+	/**
+	 * Tests writing points using the time precision feature
+	 * @throws Exception
+	 */
+	@Test
+	public void testWriteBatchWithPrecision() throws Exception {
+		// GIVEN a database and a measurement
+		String dbName = "precision_unittest_" + System.currentTimeMillis();
+		this.influxDB.createDatabase(dbName);
+
+		String rp = TestUtils.defaultRetentionPolicy(this.influxDB.version());
+
+		String measurement = TestUtils.getRandomMeasurement();
+
+		// GIVEN a batch of points using second precision
+		DateTimeFormatter formatter = DateTimeFormatter
+				.ofPattern("yyyy-MM-dd'T'HH:mm:ss'Z'")
+				.withZone(ZoneId.of("UTC"));
+		int t1 = 1485273600;
+		Point p1 = Point
+				.measurement(measurement)
+				.addField("foo", 1d)
+				.tag("device", "one")
+				.time(t1, TimeUnit.SECONDS).build(); // 2017-01-27T16:00:00
+		String timeP1 = formatter.format(Instant.ofEpochSecond(t1));
+
+		int t2 = 1485277200;
+		Point p2 = Point
+				.measurement(measurement)
+				.addField("foo", 2d)
+				.tag("device", "two")
+				.time(t2, TimeUnit.SECONDS).build(); // 2017-01-27T17:00:00
+		String timeP2 = formatter.format(Instant.ofEpochSecond(t2));
+
+		int t3 = 1485280800;
+		Point p3 = Point
+				.measurement(measurement)
+				.addField("foo", 3d)
+				.tag("device", "three")
+				.time(t3, TimeUnit.SECONDS).build(); // 2017-01-27T18:00:00
+		String timeP3 = formatter.format(Instant.ofEpochSecond(t3));
+
+		BatchPoints batchPoints = BatchPoints
+				.database(dbName)
+				.retentionPolicy(rp)
+				.precision(TimeUnit.SECONDS)
+				.points(p1, p2, p3)
+				.build();
+
+		// WHEN I write the batch
+		this.influxDB.write(batchPoints);
+
+		// THEN the measure points have a timestamp with second precision
+		QueryResult queryResult = this.influxDB.query(new Query("SELECT * FROM " + measurement, dbName));
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().size(), 3);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(0).get(0), timeP1);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(1).get(0), timeP2);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(2).get(0), timeP3);
+
+		this.influxDB.deleteDatabase(dbName);
+	}
+
+	@Test
+	public void testWriteBatchWithoutPrecision() throws Exception {
+		// GIVEN a database and a measurement
+		String dbName = "precision_unittest_" + System.currentTimeMillis();
+		this.influxDB.createDatabase(dbName);
+
+		String rp = TestUtils.defaultRetentionPolicy(this.influxDB.version());
+
+		String measurement = TestUtils.getRandomMeasurement();
+
+		// GIVEN a batch of points that has no specific precision
+		long t1 = 1485273600000000100L;
+		Point p1 = Point
+				.measurement(measurement)
+				.addField("foo", 1d)
+				.tag("device", "one")
+				.time(t1, TimeUnit.NANOSECONDS).build(); // 2017-01-27T16:00:00.000000100Z
+		Double timeP1 = Double.valueOf(t1);
+
+		long t2 = 1485277200000000200L;
+		Point p2 = Point
+				.measurement(measurement)
+				.addField("foo", 2d)
+				.tag("device", "two")
+				.time(t2, TimeUnit.NANOSECONDS).build(); // 2017-01-27T17:00:00.000000200Z
+		Double timeP2 = Double.valueOf(t2);
+
+		long t3 = 1485280800000000300L;
+		Point p3 = Point
+				.measurement(measurement)
+				.addField("foo", 3d)
+				.tag("device", "three")
+				.time(t3, TimeUnit.NANOSECONDS).build(); // 2017-01-27T18:00:00.000000300Z
+		Double timeP3 = Double.valueOf(t3);
+
+		BatchPoints batchPoints = BatchPoints
+				.database(dbName)
+				.retentionPolicy(rp)
+				.points(p1, p2, p3)
+				.build();
+
+		// WHEN I write the batch
+		this.influxDB.write(batchPoints);
+
+		// THEN the measure points have a timestamp with second precision
+		QueryResult queryResult = this.influxDB.query(new Query("SELECT * FROM " + measurement, dbName), TimeUnit.NANOSECONDS);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().size(), 3);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(0).get(0), timeP1);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(1).get(0), timeP2);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(2).get(0), timeP3);
+
+		this.influxDB.deleteDatabase(dbName);
+	}
+
+	@Test
+	public void testWriteRecordsWithPrecision() throws Exception {
+		// GIVEN a database and a measurement
+		String dbName = "precision_unittest_" + System.currentTimeMillis();
+		this.influxDB.createDatabase(dbName);
+
+		String rp = TestUtils.defaultRetentionPolicy(this.influxDB.version());
+
+		String measurement = TestUtils.getRandomMeasurement();
+
+		// GIVEN a set of records using second precision
+		DateTimeFormatter formatter = DateTimeFormatter
+				.ofPattern("yyyy-MM-dd'T'HH:mm:ss'Z'")
+				.withZone(ZoneId.of("UTC"));
+		List<String> records = new ArrayList<>();
+		records.add(measurement + ",atag=test1 idle=100,usertime=10,system=1 1485273600");
+		String timeP1 = formatter.format(Instant.ofEpochSecond(1485273600));
+
+		records.add(measurement + ",atag=test2 idle=200,usertime=20,system=2 1485277200");
+		String timeP2 = formatter.format(Instant.ofEpochSecond(1485277200));
+
+		records.add(measurement + ",atag=test3 idle=300,usertime=30,system=3 1485280800");
+		String timeP3 = formatter.format(Instant.ofEpochSecond(1485280800));
+
+		// WHEN I write the batch
+		this.influxDB.write(dbName, rp, InfluxDB.ConsistencyLevel.ONE, TimeUnit.SECONDS, records);
+
+		// THEN the measure points have a timestamp with second precision
+		QueryResult queryResult = this.influxDB.query(new Query("SELECT * FROM " + measurement, dbName));
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().size(), 3);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(0).get(0), timeP1);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(1).get(0), timeP2);
+		Assertions.assertEquals(queryResult.getResults().get(0).getSeries().get(0).getValues().get(2).get(0), timeP3);
+		this.influxDB.deleteDatabase(dbName);
+	}
 
 	/**
 	 * Test writing multiple separate records to the database using string protocol with simpler interface.