+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 2.11 [unreleased]
 
+### Features
+
+- Allow write precision of TimeUnit other than Nanoseconds [PR #321](https://github.com/influxdata/influxdb-java/pull/321)
 - Support dynamic measurement name in InfluxDBResultMapper [PR #423](https://github.com/influxdata/influxdb-java/pull/423)
 
 ## 2.10 [2018-04-26]